requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.1
django-prometheus
